Write code for adding subjects
