--- api/modules/admin/__init__.py
@@ -1,1 +1,0 @@
-from .models import Admin
