--- api/modules/admin/models/admin.py
@@ -1,24 +1,0 @@
-from api import db
-
-class Admin(db.Model):
-    __tablename__ = "admin"
-
-    id = db.Column(
-        db.Integer,
-        primary_key=True
-    )
-
-    name = db.Column(
-        db.String,
-        nullable=False
-    )
-
-    email = db.Column(
-        db.String,
-        nullable=False
-    )
-
-    password = db.Column(
-        db.String,
-        nullable=False
-    )
