--- api/modules/subject/models/subject.py
@@ -1,15 +1,0 @@
-from main import db
-
-class Subject(db.Model):
-    __tablename__ = "subject"
-
-    id = db.Column(
-        db.Integer,
-        primary_key=True
-    )
-
-    name = db.Column(
-        db.String,
-        nullable=False
-    )
-
