+++ api/requirements.txt
@@ -3,3 +3,4 @@ Flask-GraphQL==2.0.1
 Flask-Migrate==2.5.2
 Flask-SQLAlchemy==2.4.1
 graphene-sqlalchemy==2.2.2
+psycopg2-binary==2.8.4
