+++ api/schema.py
@@ -1,7 +1,7 @@
 import graphene
 from graphene import relay
 from graphene_sqlalchemy import SQLAlchemyConnectionField, SQLAlchemyObjectType
-from api.modules.admin import Admin as AdminModel
+from api.models import Admin as AdminModel, Subject as SubjectModel
 from api import db
 
 class Admin(SQLAlchemyObjectType):
@@ -27,10 +27,33 @@ class CreateAdmin(graphene.Mutation):
         return CreateAdmin(admin=admin)
 
 
+class Subject(SQLAlchemyObjectType):
+    class Meta:
+        model = SubjectModel
+        interfaces = (relay.Node, )
+
+
+class CreateSubjectInput(graphene.InputObjectType):
+    name = graphene.String(required=True)
+
+
+class CreateSubject(graphene.Mutation):
+    class Arguments:
+        subject_data = CreateSubjectInput(required=True)
+    subject = graphene.Field(Subject)
+
+    @staticmethod
+    def mutate(description, info, subject_data=None):
+        subject = SubjectModel(**subject_data)
+        db.session.add(subject)
+        db.session.commit()
+        return CreateSubject(subject=subject)
+
+
 class Viewer(graphene.ObjectType):
     all_admins = SQLAlchemyConnectionField(Admin, sort=Admin.sort_argument())
+    all_subjects = SQLAlchemyConnectionField(Subject, sort=Subject.sort_argument())
 
-    #all_vms = SQLAlchemyConnectionField(VmConnections)
 
 class Query(graphene.ObjectType):
     node = relay.Node.Field()
@@ -42,5 +65,6 @@ class Query(graphene.ObjectType):
 
 class MyMutation(graphene.ObjectType):
     create_admin = CreateAdmin.Field()
+    create_subject = CreateSubject.Field()
 
-schema = graphene.Schema(query=Query, mutation=MyMutation, types=[Admin])
+schema = graphene.Schema(query=Query, mutation=MyMutation, types=[Admin, Subject])
